Added MultAdd series and testing for dieharder's sensitivity.
Still working on evoprngs, new understandings about what makes for randomness. Simple MultiplyAdd with prime constants seems to work. The strategy of calculating a # 2x as long as desired and folding it works well. The new ways of meaningfully scaling paranoia work well. Goal is an evaluation of optimal tradeoffs of memory and processor power, keeping enough info in the cache.